Correction of return type of average function.

+++ ch5_Functions_Recursions/ch_5_Avgfor33num_fun.c
@@ -2,12 +2,12 @@
 // Create a function
 // Author : Arman
 #include <stdio.h>
-int average(int a,int b, int c);
+float average(int a,int b, int c); // function prototype
 int main(){
-printf("The average is %d", average(3,6,9) );
+printf("The average is %f", average(3,6,9) );
     return 0;
 }
 // creating a function
-int average(int a, int b, int c){
+float average(int a, int b, int c){
 return (a+b+c)/3;
 }
